[Map] Use unclipped regions + hex tile clip
This is needed because using pre-clipped regions results in the inner glow shadow being applied to the (clipped) edges of the tile for regions along the edge.

We want the shadow to follow the full (unclipped) region border off the edge of the tile (where it will be clipped).

If we don't use the full region, then when 2 tiles are next to each other, the regions along the edge will have an extra shadow along the tile edge that will look incorrect.